+++ Assets/Scripts/Sistemas/SistemaEstadisticas.cs
@@ -2,20 +2,20 @@ using UnityEngine;
 
 public class SistemaEstadisticas 
 {
-    public float vidaActual;
-    public float vidaMaxima;
-    public float vidaMinima;
+    public int vidaActual;
+    public int vidaMaxima;
+    public int vidaMinima;
 
-    public float manaActual;
-    public float manaMaximo;
-    public float manaMinimo;
+    public int manaActual;
+    public int manaMaximo;
+    public int manaMinimo;
 
     public virtual void ModificarValor(int amount) 
     {
         vidaActual = Mathf.Clamp(vidaActual + amount, vidaMinima, vidaMaxima);
     }
 
-    public virtual void RecibirDaño(float cantidad)
+    public virtual void RecibirDaño(int cantidad)
     {
         // Base implementation of damage logic
         Debug.Log($"Recibiendo {cantidad} de daño en SistemaEstadisticas.");
